requirements.txt
torch>=2.1
transformers>=4.41
accelerate>=0.30
peft>=0.11
scikit-learn>=1.3
numpy>=1.24
tqdm>=4.66
python-dotenv>=1.0
datasets>=2.18
langid==1.1.6
google-generativeai>=0.8
indic-transliteration>=2.3
matplotlib>=3.7
